# Conflicts:
#	app/src/main/java/com/zhiyicx/thinksnsplus/modules/home/message/messagecomment/MessageCommentActivity.java
#	app/src/main/java/com/zhiyicx/thinksnsplus/modules/password/changepassword/ChangePasswordContract.java
#	baseproject/src/main/java/com/zhiyicx/baseproject/impl/imageloader/glide/transformation/GlideCircleTransform.java

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/remote/UserInfoClient.java
@@ -300,7 +300,7 @@ public interface UserInfoClient {
      */
     @FormUrlEncoded
     @POST(APP_PATH_REWARD_USER)
-    Observable<Object> rewardUser(@Path("user_id") long user_id, @Field("amount") Integer amount);
+    Observable<Object> rewardUser(@Path("user_id") long user_id, @Field("amount") long amount);
 
     /*******************************************  找人  *********************************************/
